Seaparate utils module into package, providin import shim in init file
